Move difficulty icons out into its own component
