+++ src/app/components/pages/Gameboard.tsx
@@ -5,9 +5,9 @@ import {loadGameboard, logAction} from "../../state/actions";
 import * as RS from "reactstrap"
 import {Container} from "reactstrap"
 import {ShowLoading} from "../handlers/ShowLoading";
-import {Difficulty, GameboardDTO, GameboardItem, IsaacWildcard} from "../../../IsaacApiTypes";
+import {GameboardDTO, GameboardItem, IsaacWildcard} from "../../../IsaacApiTypes";
 import {TitleAndBreadcrumb} from "../elements/TitleAndBreadcrumb";
-import {difficultyShortLabelMap, NOT_FOUND, STAGE, stageLabelMap, TAG_ID, TAG_LEVEL} from "../../services/constants";
+import {NOT_FOUND, STAGE, stageLabelMap, TAG_ID, TAG_LEVEL} from "../../services/constants";
 import {isTeacher} from "../../services/user";
 import {Redirect} from "react-router";
 import {SITE, SITE_SUBJECT} from "../../services/siteConstants";
@@ -15,10 +15,9 @@ import tags from "../../services/tags";
 import {selectors} from "../../state/selectors";
 import {showWildcard} from "../../services/gameboards";
 import queryString from "query-string";
-import {calculateHexagonProportions, Hexagon} from "../elements/svg/Hexagon";
-import {Rectangle} from "../elements/svg/Rectangle";
 import {determineAudienceViews} from "../../services/userContext";
 import classnames from "classnames";
+import {DifficultyIcons} from "../elements/svg/DifficultyIcons";
 
 function extractFilterQueryString(gameboard: GameboardDTO): string {
     const csvQuery: {[key: string]: string} = {}
@@ -37,39 +36,6 @@ function getTags(docTags?: string[]) {
     if (!docTags) return [];
 
     return tags.getByIdsAsHierarchy(docTags as TAG_ID[]);
-}
-
-const DifficultyIcons = ({difficulty} : {difficulty : Difficulty}) => {
-    // Difficulty icon proportions
-    const difficultyIconWidth = 25;
-    const difficultyIconXPadding = 1.5;
-    const yPadding = 2;
-    const miniHexagon = calculateHexagonProportions(difficultyIconWidth / 2, 0);
-    const miniSquare = {width: difficultyIconWidth, height: difficultyIconWidth};
-
-    const difficultyLabel = difficultyShortLabelMap[difficulty];
-    const difficultyCategory = difficultyLabel[0];
-    const numberOfLevelsForDifficultyCategory = 3;
-    const difficultyLevel = parseInt(difficultyLabel[1]);
-
-    return <div aria-label={"Difficulty: " + difficultyLabel}>
-        <svg width={`${numberOfLevelsForDifficultyCategory * (difficultyIconWidth + 2 * difficultyIconXPadding) - difficultyIconXPadding}px`} height={`${miniHexagon.quarterHeight * 4 + 2 * yPadding}px`} >
-            {Array(numberOfLevelsForDifficultyCategory).fill(undefined).map((_, i) => {
-                const active = i < difficultyLevel;
-                return <g key={i} transform={`translate(${i * (difficultyIconWidth + 2 * difficultyIconXPadding)}, ${yPadding + (difficultyCategory === "P" ? 0 : 2)})`}>
-                    {difficultyCategory === "P" ?
-                        <Hexagon {...miniHexagon} className={"hex difficulty practice " + classnames({active})} /> :
-                        <Rectangle {...miniSquare} className={"square difficulty challenge " + classnames({active})} />
-                    }
-                    {<foreignObject width={difficultyIconWidth} height={difficultyIconWidth + (difficultyCategory === "P" ? yPadding + 2 : 0)}>
-                        <div className={`difficulty-title ${classnames({active})} difficulty-${i + 1}`}>
-                            {difficultyCategory}
-                        </div>
-                    </foreignObject>}
-                </g>;
-            })}
-        </svg>
-    </div>
 }
 
 const GameboardItemComponent = ({gameboard, question}: {gameboard: GameboardDTO, question: GameboardItem}) => {
